refactor: Git同期を src/koda/git_sync.py に切り出す (#29)
- GitSyncPayload: JSONL dump/load/parse_record
- GitSyncRepo: git CLI ラッパー (ensure_worktree/has_remote/
  preferred_remote/current_branch/has_upstream/pull_rebase_if_remote/
  push_if_remote)
- MemoMerger: リモート JSONL を local DB に取り込む状態機械 (merge)
- require_git_cli / require_jsonl_format / resolve_sync_root /
  resolve_payload_path / atomic_write_bytes / parse_memo_datetime も移動
- main.py の push / pull は薄いディスパッチに簡略化
- upstream 未設定でも push/pull が動作する挙動は維持 (783534a)
